Fix amd64 issues as reported on bug #311158. Original patch from Ubuntu or
2006-01-10  Peter Johanson <peter@peterjohanson.com>
       * plugins/egg-sharp.cs: Fix amd64 issues as reported on bug #311158.
       Original patch from Ubuntu or wherever they got the patch from.

+++ plugins/egg-sharp.cs
@@ -185,7 +185,7 @@ namespace Egg
 	[StructLayout(LayoutKind.Sequential)]
 	internal struct XClientMessageEvent {
 		internal XEventName     type;
-		internal int            serial;
+		internal IntPtr		serial;
 		internal bool           send_event;
 		internal IntPtr         display;
 		internal IntPtr         window;
